Renamed get_parents.py to utils.py and changed return value of cli.py functions to workflow
