+++ trampolino/utils.py
